+++ src/ast.rs
@@ -10,7 +10,6 @@
  * Most clients will be interested in visiting the nodes of an ast.
  */
 use crate::{
-    common::unescape_string,
     flog::{flog, flogf},
     parse_constants::{
         ERROR_BAD_COMMAND_ASSIGN_ERR_MSG, INVALID_PIPELINE_CMD_ERR_MSG, ParseError, ParseErrorCode,
@@ -24,7 +23,7 @@ use crate::{
         TokFlags, TokenType, Tokenizer, TokenizerError, variable_assignment_equals_pos,
     },
 };
-use fish_common::UnescapeStringStyle;
+use fish_common::{UnescapeStringStyle, unescape_string};
 use macro_rules_attribute::derive;
 use std::borrow::Cow;
 use std::convert::AsMut;
