+++ src/bin/fish.rs
@@ -24,7 +24,7 @@ use fish::{
         fish_indent, fish_key_reader,
         shared::{STATUS_CMD_ERROR, STATUS_CMD_OK, STATUS_CMD_UNKNOWN, VERSION_STRING_TEMPLATE},
     },
-    common::{PACKAGE_NAME, PROFILING_ACTIVE, PROGRAM_NAME, escape},
+    common::{PACKAGE_NAME, PROFILING_ACTIVE, PROGRAM_NAME},
     env::{
         EnvMode, Statuses,
         config_paths::ConfigPaths,
@@ -56,7 +56,7 @@ use fish::{
     topic_monitor,
     wutil::waccess,
 };
-use fish_common::save_term_foreground_process_group;
+use fish_common::{escape, save_term_foreground_process_group};
 use fish_widestring::{bytes2wcstring, osstr2wcstring, wcs2bytes};
 use libc::{STDERR_FILENO, STDIN_FILENO};
 use nix::{
