+++ src/builtins/source.rs
@@ -1,9 +1,9 @@
 use super::prelude::*;
 use crate::{
-    builtins::error::Error, common::escape, err_fmt, err_raw, err_str, fds::wopen_cloexec,
-    nix::isatty, parser::Block, reader::reader_read,
+    builtins::error::Error, err_fmt, err_raw, err_str, fds::wopen_cloexec, nix::isatty,
+    parser::Block, reader::reader_read,
 };
-use fish_common::FilenameRef;
+use fish_common::{FilenameRef, escape};
 use nix::{fcntl::OFlag, sys::stat::Mode};
 use std::os::fd::AsRawFd as _;
